+++ src/app/modules/modals/post-register-modal/post-register-modal.component.ts
@@ -1,0 +1,17 @@
+import { Component } from '@angular/core';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+
+@Component({
+  selector: 'app-post-register-modal',
+  templateUrl: './post-register-modal.component.html',
+  styleUrl: './post-register-modal.component.scss'
+})
+export class PostRegisterModalComponent {
+
+  constructor(private modalService: NgbModal){
+  }
+
+  closeModal() {
+    this.modalService.dismissAll();
+  }
+}